Use DTOs and prepare for code reduction

+++ src/main/java/com/javainiaisuzspringom/tripperis/repositories/ApartmentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ApartmentRepository extends JpaRepository<Apartment, Long> {
+public interface ApartmentRepository extends JpaRepository<Apartment, Integer> {
 
 }
